+++ src/application.rs
@@ -1,4 +1,4 @@
-pub mod commands;
 pub mod common;
+pub mod interactors;
 pub mod set;
 pub mod user;
