+++ src/application/interactors/set_deleted_col.rs
