+++ src/middlewares.rs
@@ -1,9 +1,7 @@
 mod client_application;
 mod create_user;
 mod database;
-mod deleted_sets;
 
 pub use client_application::{Client, ClientApplicationMiddleware};
 pub use create_user::CreateUserMiddleware;
 pub use database::DatabaseMiddleware;
-pub use deleted_sets::DeletedSetsMiddleware;
